Add community data to stats view
