+++ app/src/marketplace/components/ProductPage/CommunityPending/communityPending.pcss
@@ -4,7 +4,6 @@
   grid-column-gap: 4em;
   align-items: center;
   justify-content: center;
-  padding-top: 4em;
 
   .deployMessageHeading {
     font-size: 30px;
