+++ app/src/marketplace/containers/ProductPage/communityStats.pcss
@@ -11,6 +11,10 @@
   grid-template-rows: 4.5em auto auto 4.5em;
 }
 
+.communityPending {
+  padding-top: 4em;
+}
+
 .stats {
   padding: 2rem;
 }
