+++ app/src/userpages/components/ProductsPage/Stats/stats.pcss
@@ -1,3 +1,9 @@
-.searchPlaceholder {
-  width: var(--um);
+.loadingIndicator {
+  top: 2px;
+}
+
+.statBox {
+  background: white;
+  border-radius: 4px;
+  padding: 2.5rem;
 }
